+++ src-dbg/flow_storm/debugger/ui/value_inspector.clj
@@ -14,7 +14,6 @@
            [javafx.scene.control SplitPane]
            [javafx.scene.layout VBox HBox Priority]))
 
-(set! *warn-on-reflection* true)
 
 (declare create-value-pane)