+++ src-dbg/flow_storm/debugger/ui/value_renderers.clj
@@ -9,7 +9,6 @@
            [javafx.scene Node]
            [javafx.scene.input MouseEvent]))
 
-(set! *warn-on-reflection* true)
 
 ;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;;
 ;; NOTE :                                                                                              ;;