+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 	
 ### New Features
 
+    - Implement storm instrumentation management in the browser
     - Add "Copy qualified function symbol" and "Copy function calling form" to form menu
     
 ### Changes